chore: Update tb_delivery and tb_delivery_items tables
This commit updates the `tb_delivery` and `tb_delivery_items` tables in the `bill.sql` file. It adds the `delivery_weight_total` column to the `tb_delivery` table and the `item_weight` column to the `tb_delivery_items` table. These columns are used to store the weight information for deliveries and items, respectively.